reorg parts into cmd/gnatsd to separate gnatsd as library and application
