+++ auth/plain.go
@@ -1,7 +1,7 @@
 package auth
 
 import (
-	"github.com/nats-io/gnatsd/server"
+	"github.com/nats-io/gnatsd"
 )
 
 type Plain struct {
@@ -9,7 +9,7 @@ type Plain struct {
 	Password string
 }
 
-func (p *Plain) Check(c server.ClientAuth) bool {
+func (p *Plain) Check(c gnatsd.ClientAuth) bool {
 	opts := c.GetOpts()
 	if p.Username != opts.Username || p.Password != opts.Password {
 		return false
