+++ auth/token.go
@@ -1,14 +1,14 @@
 package auth
 
 import (
-	"github.com/nats-io/gnatsd/server"
+	"github.com/nats-io/gnatsd"
 )
 
 type Token struct {
 	Token string
 }
 
-func (p *Token) Check(c server.ClientAuth) bool {
+func (p *Token) Check(c gnatsd.ClientAuth) bool {
 	opts := c.GetOpts()
 	if p.Token != opts.Authorization {
 		return false
