+++ log.go
@@ -1,6 +1,6 @@
 // Copyright 2012-2015 Apcera Inc. All rights reserved.
 
-package server
+package gnatsd
 
 import (
 	"sync"
