+++ parser.go
@@ -1,6 +1,6 @@
 // Copyright 2012-2014 Apcera Inc. All rights reserved.
 
-package server
+package gnatsd
 
 import (
 	"fmt"
