+++ pse_freebsd.go
@@ -1,8 +1,6 @@
 // Copyright 2015 Apcera Inc. All rights reserved.
 
-package server
-
-import ()
+package gnatsd
 
 // This is a placeholder for now.
 func procUsage(pcpu *float64, rss, vss *int64) error {
